Moving aux methods to their own file for separtion of processing logic / methods
